mvp_pointax_huvfa/requirements.txt
jax>=0.4.13
jaxlib>=0.4.13
flax>=0.8.0
optax>=0.1.8
chex>=0.1.86
PyYAML>=6.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
